+++ Source/Lib/Morris.Moxy/Extensions/AttributeSyntaxGetArgumentKeyValuePairsExtension.cs
@@ -1,5 +1,7 @@
 ﻿using Microsoft.CodeAnalysis;
+using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
+using Morris.Moxy.Metas.ScriptVariables;
 using System.Collections.Immutable;
 using System.Runtime.CompilerServices;
 
@@ -8,7 +10,7 @@ namespace Morris.Moxy.Extensions;
 internal static class AttributeSyntaxGetArgumentKeyValuePairsExtension
 {
 	[MethodImpl(MethodImplOptions.AggressiveInlining)]
-	public static ImmutableArray<KeyValuePair<string, string>> GetArgumentKeyValuePairs(
+	public static ImmutableArray<KeyValuePair<string, object?>> GetArgumentKeyValuePairs(
 		this AttributeSyntax attributeSyntax,
 		SemanticModel semanticModel,
 		ImmutableArray<Metas.Templates.TemplateInput> requiredInputs,
@@ -16,13 +18,15 @@ internal static class AttributeSyntaxGetArgumentKeyValuePairsExtension
 	{
 		SeparatedSyntaxList<AttributeArgumentSyntax>? arguments = attributeSyntax.ArgumentList?.Arguments;
 		if (arguments is null)
-			return ImmutableArray<KeyValuePair<string, string>>.Empty;
+			return ImmutableArray<KeyValuePair<string, object?>>.Empty;
 
 		ImmutableArray<Metas.Templates.TemplateInput> allInputs = requiredInputs.AddRange(optionalInputs);
 
-		var resultBuilder = ImmutableArray.CreateBuilder<KeyValuePair<string, string>>();
+		var nameToValueLookup = new Dictionary<string, object?>();
+		foreach (var item in optionalInputs.Union(requiredInputs).Where(x => x.DefaultValue is not null))
+			nameToValueLookup[item.Name] = GetValueFromStringRepresentation(semanticModel, item.DefaultValue!);
 
-		for(int argumentIndex = 0; argumentIndex < arguments.Value.Count; argumentIndex++)
+		for (int argumentIndex = 0; argumentIndex < arguments.Value.Count; argumentIndex++)
 		{
 			AttributeArgumentSyntax argument = arguments.Value[argumentIndex];
 
@@ -31,23 +35,55 @@ internal static class AttributeSyntaxGetArgumentKeyValuePairsExtension
 				? argument.NameEquals.Name.Identifier.ValueText
 				: argument.NameColon is not null
 				? argument.NameColon.Name.Identifier.ValueText
-				: allInputs[argumentIndex].Name;
+				: allInputs.Length > argumentIndex
+				? allInputs[argumentIndex].Name
+				: "";
 
-			string value = argument.Expression switch {
-				TypeOfExpressionSyntax x => x.ToFullString(),
-				_ => TrimQuotes(argument)
-			};
+			object? value = GetValueFromStringRepresentation(semanticModel, argument);
+			nameToValueLookup[argumentName] = value;
 
-			resultBuilder.Add(new KeyValuePair<string, string>(argumentName, value));
 		}
-
-		return resultBuilder.ToImmutableArray();
+		return nameToValueLookup.ToImmutableArray();
 	}
 
 
 	[MethodImpl(MethodImplOptions.AggressiveInlining)]
-	private static string TrimQuotes(AttributeArgumentSyntax argument) =>
-		argument.Expression.ToFullString() switch {
+	private static object? GetValueFromStringRepresentation(SemanticModel model, string expressionStr)
+	{
+		var expression = SyntaxFactory.ParseExpression(expressionStr);
+		return GetValueFromArgument(model, expression);
+	}
+
+	[MethodImpl(MethodImplOptions.AggressiveInlining)]
+	private static object? GetValueFromStringRepresentation(SemanticModel model, AttributeArgumentSyntax syntax)
+	{
+		var expression = syntax.Expression;
+		return GetValueFromArgument(model, expression);
+	}
+
+	[MethodImpl(MethodImplOptions.AggressiveInlining)]
+	private static object? GetValueFromArgument(SemanticModel model, ExpressionSyntax expression)
+	{
+		return expression switch {
+			LiteralExpressionSyntax lit when lit.Token.IsKind(SyntaxKind.TrueKeyword) => true,
+			LiteralExpressionSyntax lit when lit.Token.IsKind(SyntaxKind.FalseKeyword) => false,
+			LiteralExpressionSyntax lit when lit.Token.IsKind(SyntaxKind.NullKeyword) => null,
+			TypeOfExpressionSyntax x => CreateTypeVariable(model, x),
+			_ => TrimQuotes(expression)
+		};
+	}
+
+	private static TypeVariable CreateTypeVariable(SemanticModel model, TypeOfExpressionSyntax x)
+	{
+		ITypeSymbol modelType = model.GetTypeInfo(x.Type).Type!;
+		return new TypeVariable(
+			name: modelType.Name,
+			fullName: modelType.ToDisplayString());
+	}
+
+	[MethodImpl(MethodImplOptions.AggressiveInlining)]
+	private static string TrimQuotes(ExpressionSyntax expression) =>
+		expression.ToFullString() switch {
 			string x when x.StartsWith("\"") => x.Substring(1, x.Length - 2),
 			string x => x,
 			_ => throw new NotImplementedException()