+++ Source/Lib/Morris.Moxy/SourceGenerators/TemplateAttributeSourceGenerator.cs
@@ -36,7 +36,7 @@ internal static class TemplateAttributeSourceGenerator
 			productionContext.AddCompilationError("", compilationError);
 		}
 
-		if (generatedSourceCode is not null)
+		if (generatedSourceCode is not null && !productionContext.CancellationToken.IsCancellationRequested)
 			productionContext.AddSource(
 				hintName: classFileName,
 				source: generatedSourceCode);
@@ -75,8 +75,8 @@ internal static class TemplateAttributeSourceGenerator
 			writer.WriteLine($"using {attributeUsingClause};");
 		writer.WriteLine();
 
-		writer.WriteLine("[AttributeUsage(AttributeTargets.Class | AttributeTargets.Struct, AllowMultiple = true)]");
-		writer.WriteLine($"internal class {parsedTemplate.Name}Attribute : Attribute");
+		writer.WriteLine("[System.AttributeUsage(System.AttributeTargets.Class | System.AttributeTargets.Struct, AllowMultiple = true)]");
+		writer.WriteLine($"internal class {parsedTemplate.Name}Attribute : System.Attribute");
 		using (writer.CodeBlock())
 		{
 			GenerateClassProperties(parsedTemplate, writer);