+++ Source/Tutorials/01-MixInAMethod/MixInAMethod/Person.cs
@@ -1,6 +1,11 @@
 ﻿namespace MixInAMethod;
 
+[ShowMoxyVariables]
 [SayHello]
 internal partial class Person
 {
+	public void Ha()
+	{
+		Console.WriteLine("Ha");
+	}
 }