+++ Source/Tutorials/01-MixInAMethod/MixInAMethod/Program.cs
@@ -2,7 +2,10 @@
 
 using MixInAMethod;
 
-Console.WriteLine("Hello, World!");
+System.Console.WriteLine("Hello, World!");
 
 var person = new Person();
 person.SayHello();
+person.ShowMoxyVariables();
+var nested = new Container.Nested();
+nested.ShowMoxyVariables();